competitive-programming-agents/requirements.txt
jinja2>=3.1.0
litellm>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.64.0 